refactor Session
