--- packages/space-repetition/src/types/Session.ts
@@ -1,23 +1,0 @@
-import type { Attempt } from "./Attempt";
-
-export type Session = {
-  sessionId: string;
-  attempts: Attempt[];
-};
-export const Session = (attempts: Attempt[]): Session => {
-  const sessionId = attempts.at(0)?.reviewSessionId;
-  if (!sessionId || !attempts.every((a) => a.reviewSessionId === sessionId)) {
-    throw new Error("All attempts must have the same reviewSessionId");
-  }
-
-  return {
-    sessionId: sessionId,
-    attempts,
-  };
-};
-export type SessionResult = {
-  sessionId: string;
-  attempts: Attempt[];
-  correct: number;
-  total: number;
-};
